handling partitionMetadataStore during updates
