+++ astra/src/main/java/com/slack/astra/metadata/partition/PartitionMetadata.java
@@ -3,11 +3,14 @@ package com.slack.astra.metadata.partition;
 import com.slack.astra.metadata.core.AstraMetadata;
 import java.util.Objects;
 
-/** PartitionMetadata Object to track the utilization and isPartitionShared in zookeeper */
+/**
+ * PartitionMetadata Object to track the utilization and isPartitionShared (set to true when
+ * partition is shared between Tenants) in zookeeper
+ */
 public class PartitionMetadata extends AstraMetadata {
   public final String partitionId;
-  public final long utilization;
-  public final boolean isPartitionShared;
+  public long utilization;
+  public boolean isPartitionShared;
 
   public PartitionMetadata(String partition, long utilization, boolean isPartitionShared) {
     super(partition);
